Updated Source Code to v4.3
+Added sound detection (audio peak measurement)
+Added Option to cancel shutdown while sound is playing
+Added debug window to show detailed information
+Implemented Log System
+Fixed some bugs
+Code improvements and code commenting
+Improved exception handling

+++ SourceCode/TimeoutWindow.h
@@ -19,7 +19,7 @@ namespace StandBye {
 	public ref class TimeoutWindow : public MetroFramework::Forms::MetroForm
 	{
 		int delay; // in milliseconds
-		double startTime; //in milliseconds
+		int startTime; //in milliseconds
 		System::Windows::Forms::Timer^ timer;
 		SettingsProvider* settings_provider;
 		int counter;
